+++ axelor-business-production/src/main/java/com/axelor/apps/businessproduction/service/InvoicingProjectServiceBusinessProdImpl.java
@@ -42,6 +42,11 @@ public class InvoicingProjectServiceBusinessProdImpl extends InvoicingProjectSer
     counter++;
 
     this.fillLines(invoicingProject, project);
+
+    if (!invoicingProject.getConsolidatePhaseWhenInvoicing()) {
+      return;
+    }
+
     List<Project> projectChildrenList =
         Beans.get(ProjectRepository.class).all().filter("self.parentProject = ?1", project).fetch();